fix: resolve all TypeScript errors across backend (87->0) and fix Soroban SDK contract compatibility

+++ contracts/src/proctoring.rs
@@ -1,4 +1,3 @@
-#![no_std]
 use soroban_sdk::{
     contract, contractimpl, contracttype, symbol_short, Address, Env, String, Vec, BytesN,
 };
@@ -22,7 +21,7 @@ pub struct AssessmentSession {
     pub start_time: u64,
     pub end_time: Option<u64>,
     pub identity_hash: BytesN<32>, // Biometric data hash
-    pub status: u8, // 0: Pending, 1: Active, 2: Completed, 3: Flagged, 4: Invalidated
+    pub status: u32, // 0: Pending, 1: Active, 2: Completed, 3: Flagged, 4: Invalidated
 }
 
 #[contracttype]
